requirements.txt
asgiref==3.4.1
Django==4.0
gunicorn==20.1.0
pkg_resources==0.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
sqlparse==0.4.2
djangorestframework==3.13.1
beautifulsoup4==4.10.0
requests==2.27.0
lxml==4.7.1
selenium==4.1.0
webdriver-manager
python-dotenv==0.19.2
celery==5.2.3
kombu